+++ tests/test_adapter.py
@@ -158,7 +158,9 @@ def test_inf_timestamps(connection: HarlequinPostgresConnection) -> None:
     ]
 
 
-def test_closed_conn_raises_right_error(connection: HarlequinPostgresConnection) -> None:
+def test_closed_conn_raises_right_error(
+    connection: HarlequinPostgresConnection,
+) -> None:
     connection._main_conn.close()
 
     with pytest.raises(HarlequinQueryError):